fix and clean customer domain

+++ database/migrations/2026_06_28_185845_create_customer_profiles_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('company_id')->nullable();
             $table->foreign('company_id')->references('id')->on('companies');
             $table->string('name');
-            $table->integer('type');
+            $table->unsignedTinyInteger('type');
             $table->boolean('is_deleted')->default(false);
             $table->timestamps();
         });
